+++ src/main/java/com/github/sitture/envconfig/EnvironmentVariables.java
@@ -1,4 +1,4 @@
-package com.github.sitture.env.config;
+package com.github.sitture.envconfig;
 
 import org.apache.commons.configuration2.Configuration;
 import org.apache.commons.configuration2.MapConfiguration;
@@ -6,8 +6,6 @@ import org.apache.commons.configuration2.SystemConfiguration;
 
 import java.util.HashMap;
 import java.util.Map;
-
-import static com.github.sitture.env.config.utils.PropertyUtils.getProcessedEnvKey;
 
 class EnvironmentVariables {
 
@@ -19,21 +17,21 @@ class EnvironmentVariables {
 		environmentConfiguration = new MapConfiguration(getEnvMap());
 	}
 
+	private static Map<String, String> getEnvMap() {
+		final Map<String, String> envMap = new HashMap<>();
+		System.getenv().forEach((key, value) -> {
+			envMap.put(key, value);
+			envMap.put(EnvConfigUtils.getProcessedEnvKey(key), value);
+		});
+		return envMap;
+	}
+
 	protected Configuration getSystemConfiguration() {
 		return systemConfiguration;
 	}
 
 	protected Configuration getEnvironmentConfiguration() {
 		return environmentConfiguration;
-	}
-
-	private static Map<String, String> getEnvMap() {
-		final Map<String, String> envMap = new HashMap<>();
-		System.getenv().entrySet().forEach(entry -> {
-			envMap.put(entry.getKey(), entry.getValue());
-			envMap.put(getProcessedEnvKey(entry.getKey()), entry.getValue());
-		});
-		return envMap;
 	}
 
 }