+++ src/test/java/com/github/sitture/envconfig/EnvConfigTest.java
@@ -1,4 +1,4 @@
-package com.github.sitture.env.config;
+package com.github.sitture.envconfig;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,12 +14,8 @@ import java.util.List;
 @ExtendWith(SystemStubsExtension.class)
 class EnvConfigTest {
 
-	private static final String CONFIG_ENV_KEY = "config.env";
-	private static final String CONFIG_KEEPASS_ENABLED_KEY = "config.keepass.enabled";
-	private static final String CONFIG_KEEPASS_MASTERKEY_KEY = "config.keepass.masterkey";
 	private static final String CONFIG_KEEPASS_PASSWORD = "envconfig";
 	private static final String TEST_ENVIRONMENT = "test";
-	private static final String DEFAULT_ENVIRONMENT = "default";
 	private static final String TEST_PROPERTY = "property";
 	private static final String TEST_VALUE = "value";
 	private static final String KEEPASS_VALUE = "KEEPASS_VALUE";
@@ -29,7 +25,9 @@ class EnvConfigTest {
 
 	@BeforeEach
 	void setUp() {
-		System.clearProperty("config.keepass.filename");
+		System.clearProperty(EnvConfigProperties.CONFIG_KEEPASS_FILENAME_KEY);
+		System.clearProperty(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY);
+		System.clearProperty(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY);
 		EnvConfig.reset();
 	}
 
@@ -40,29 +38,29 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetDefaultEnvironment() {
-		System.setProperty(CONFIG_ENV_KEY, DEFAULT_ENVIRONMENT);
-		Assertions.assertEquals(DEFAULT_ENVIRONMENT, EnvConfig.getEnvironment());
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, EnvConfigProperties.CONFIG_ENV_DEFAULT);
+		Assertions.assertEquals(EnvConfigProperties.CONFIG_ENV_DEFAULT, EnvConfig.getEnvironment());
 		// when env not set
-		System.clearProperty(CONFIG_ENV_KEY);
-		Assertions.assertEquals(DEFAULT_ENVIRONMENT, EnvConfig.getEnvironment());
+		System.clearProperty(EnvConfigProperties.CONFIG_ENV_KEY);
+		Assertions.assertEquals(EnvConfigProperties.CONFIG_ENV_DEFAULT, EnvConfig.getEnvironment());
 	}
 
 	@Test
 	void testCanGetEnvironment() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		Assertions.assertEquals(TEST_ENVIRONMENT, EnvConfig.getEnvironment());
 	}
 
 	@Test
 	void testCanGetEnvironmentWhenMultiple() {
-		System.setProperty(CONFIG_ENV_KEY, "default , test");
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "default , test");
 		Assertions.assertEquals(TEST_ENVIRONMENT, EnvConfig.getEnvironment());
 	}
 
 	@Test
 	void testExceptionWhenEnvMissing() {
-		System.setProperty(CONFIG_ENV_KEY, "default, non-existing ");
-		final ConfigException exception = Assertions.assertThrows(ConfigException.class,
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "default, non-existing ");
+		final EnvConfigException exception = Assertions.assertThrows(EnvConfigException.class,
 				EnvConfig::getEnvironment);
 		Assertions.assertTrue(exception.getMessage()
 				.endsWith("/env-config/config/non-existing' does not exist or not a valid config directory!"));
@@ -70,32 +68,53 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetEnvironmentWithASpace() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT + " ");
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT + " ");
 		Assertions.assertEquals(TEST_ENVIRONMENT, EnvConfig.getEnvironment());
 	}
 
 	@Test
+	void testThrowsExceptionWhenNoPropertiesInEnv() {
+		// given env is default and empty-profile exists in env properties
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "empty-env");
+		// then an exception is thrown
+		final EnvConfigException exception = Assertions.assertThrows(EnvConfigException.class,
+				() -> EnvConfig.getOrThrow("non.existing"));
+		Assertions.assertTrue(exception.getMessage().startsWith("No property files found under"), exception.getMessage());
+		Assertions.assertTrue(exception.getMessage().endsWith("/env-config/config/empty-env'"), exception.getMessage());
+	}
+
+	@Test
 	void testCanGetProperty() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		Assertions.assertEquals("my_value", EnvConfig.get("my.property"));
 	}
 
 	@Test
+	void testCanGetPropertyUsingBothFormats() {
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		// when PROPERTY_SEVEN exists in test environment
+		// then should be able to get key using both properties and env var formats
+
+		Assertions.assertEquals("test", EnvConfig.get("PROPERTY_SEVEN"));
+		Assertions.assertEquals("test", EnvConfig.get("property.seven"));
+	}
+
+	@Test
 	void testCanGetPropertyFromGetOrThrow() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		Assertions.assertEquals("my_value", EnvConfig.getOrThrow("my.property"));
 	}
 
 	@Test
 	void testDoesNotGetsPropertyFromSubDirs() {
-		System.setProperty(CONFIG_ENV_KEY, DEFAULT_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, EnvConfigProperties.CONFIG_ENV_DEFAULT);
 		Assertions.assertNull(EnvConfig.get("property.sub.dir"));
 	}
 
 	@Test
 	void testCanGetPropertyWhenMultipleEnv() {
 		final String testEnv = "test-env";
-		System.setProperty(CONFIG_ENV_KEY, "test," + testEnv);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "test," + testEnv);
 		Assertions.assertEquals(testEnv, EnvConfig.getEnvironment());
 		// When a property.one exists in all environments, including default
 		Assertions.assertEquals(testEnv, EnvConfig.get("property.one"));
@@ -104,14 +123,14 @@ class EnvConfigTest {
 		// When a property.three does not exist in current environment
 		Assertions.assertEquals("test", EnvConfig.get("property.three"));
 		// When a property.four exists in default environment only
-		Assertions.assertEquals(DEFAULT_ENVIRONMENT, EnvConfig.get("property.four"));
+		Assertions.assertEquals(EnvConfigProperties.CONFIG_ENV_DEFAULT, EnvConfig.get("property.four"));
 	}
 
 	@Test
 	void testCanGetEntryWhenEnvVarAndDefaultValueSameWithParentEnv() {
 		environmentVariables.set("property.five", "default");
 		final String testEnv = "test-env";
-		System.setProperty(CONFIG_ENV_KEY, "test," + testEnv);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "test," + testEnv);
 		// when property.five is set as env variable
 		// and does not exist in test-env
 		// and exists in default env with same value as env var
@@ -124,19 +143,20 @@ class EnvConfigTest {
 	void testCanGetEntryWhenEnvVarAndDefaultValueSameWithSingleEnv() {
 		final var key = "PROPERTY_SEVEN";
 		environmentVariables.set(key, "default");
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		// when PROPERTY_SEVEN=default is set as env variable
 		// and exists in default env with same value as env var
 		// and exists in test env with PROPERTY_SEVEN=test
 		// then value from env variable takes priority
 		Assertions.assertEquals("default", EnvConfig.get(key));
+		Assertions.assertEquals("default", EnvConfig.get(EnvConfigUtils.getProcessedEnvKey(key)));
 	}
 
 	@Test
 	void testCanGetEntryWhenEnvVarAndDefaultValueDifferent() {
 		environmentVariables.set("property.five", "env.default");
 		final String testEnv = "test-env";
-		System.setProperty(CONFIG_ENV_KEY, "test," + testEnv);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "test," + testEnv);
 		// when property.five is set as env variable
 		// and does not exist in test-env
 		// and exists in test env
@@ -149,7 +169,7 @@ class EnvConfigTest {
 	void testCanGetEntryWhenEnvVarSet() {
 		environmentVariables.set("property.six", "env.property.six");
 		final String testEnv = "test-env";
-		System.setProperty(CONFIG_ENV_KEY, "test," + testEnv);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "test," + testEnv);
 		// when property.six is set as env variable
 		// and exists in test-env env
 		// and exists in test env
@@ -161,7 +181,7 @@ class EnvConfigTest {
 	@Test
 	void testCanGetEntryWhenEnvVarSetInMultiEnvs() {
 		final String testEnv = "test-env";
-		System.setProperty(CONFIG_ENV_KEY, "test," + testEnv);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "test," + testEnv);
 		// when property.seven is not set in test-env
 		// and exists in test env in environment variable format. i.e. PROPERTY_SEVEN=test
 		// and exists in default env with different value i.e. PROPERTY_SEVEN=default
@@ -171,7 +191,7 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetPropertyFromEnvVars() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		environmentVariables.set("MY_ENV_PROPERTY", "my_env_value");
 		Assertions.assertEquals("my_env_value", EnvConfig.get("my.env.property"));
 		Assertions.assertEquals("my_env_value", EnvConfig.get("MY_ENV_PROPERTY"));
@@ -182,7 +202,7 @@ class EnvConfigTest {
 	@Test
 	void testCanGetParsedInt() {
 		System.setProperty(TEST_PROPERTY, "123");
-		Assertions.assertEquals(123, EnvConfig.getInteger(TEST_PROPERTY));
+		Assertions.assertEquals(123, EnvConfig.getInt(TEST_PROPERTY));
 	}
 
 	@Test
@@ -197,7 +217,7 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetParsedList() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		Assertions.assertTrue(EnvConfig.getList(TEST_PROPERTY).isEmpty());
 		System.setProperty(TEST_PROPERTY, "env");
 		Assertions.assertEquals(1, EnvConfig.getList(TEST_PROPERTY).size());
@@ -224,7 +244,7 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetDefaultForNonExistingProperty() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		Assertions.assertEquals("test", EnvConfig.get("non.existing", "test"));
 	}
 
@@ -235,13 +255,13 @@ class EnvConfigTest {
 
 	@Test
 	void testNullForNonExistingProperty() {
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		Assertions.assertNull(EnvConfig.get("non.existing"));
 	}
 
 	@Test
 	void testMissingVariableExceptionThrown() {
-		final MissingVariableException exception = Assertions.assertThrows(MissingVariableException.class,
+		final EnvConfigException exception = Assertions.assertThrows(EnvConfigException.class,
 				() -> EnvConfig.getOrThrow("non.existing"));
 		Assertions.assertEquals("Missing required variable 'non.existing'", exception.getMessage());
 	}
@@ -249,11 +269,11 @@ class EnvConfigTest {
 	@Test
 	void testCanAddANewProperty() {
 		EnvConfig.add("1", 1);
-		Assertions.assertEquals(1, EnvConfig.getInteger("1"));
+		Assertions.assertEquals(1, EnvConfig.getInt("1"));
 		Assertions.assertEquals("1", EnvConfig.get("1"));
 		Assertions.assertFalse(EnvConfig.getBool("1"));
 		EnvConfig.add("1", "1");
-		Assertions.assertEquals(1, EnvConfig.getInteger("1"));
+		Assertions.assertEquals(1, EnvConfig.getInt("1"));
 		Assertions.assertEquals("1", EnvConfig.get("1"));
 		EnvConfig.add("flag", true);
 		Assertions.assertTrue(EnvConfig.getBool("flag"));
@@ -278,10 +298,41 @@ class EnvConfigTest {
 		Assertions.assertNull(System.getProperty(TEST_PROPERTY));
 	}
 
+	@Test
+	void testCanGetEntryFromKeepassWhenFileNameSpecified() {
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_FILENAME_KEY, "env-config.kdbx");
+		enabledKeepass();
+		Assertions.assertEquals(KEEPASS_VALUE, EnvConfig.get("my.keepass.property"));
+	}
+
 	private void enabledKeepass() {
-		System.setProperty(CONFIG_KEEPASS_ENABLED_KEY, "true");
-		System.setProperty(CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
-		System.setProperty(CONFIG_ENV_KEY, DEFAULT_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY, "true");
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, EnvConfigProperties.CONFIG_ENV_DEFAULT);
+	}
+
+	@Test
+	void testCanGetFromKeepassWhenFileNameWithSpace() {
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_FILENAME_KEY, "env config.kdbx");
+		enabledKeepass();
+		Assertions.assertEquals(KEEPASS_VALUE, EnvConfig.get("my.keepass.property"));
+	}
+
+	@Test
+	void testExceptionWhenKeepassFileMissing() {
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_FILENAME_KEY, "non-existing");
+		final EnvConfigException exception = Assertions.assertThrows(EnvConfigException.class,
+				this::testCanGetEntryFromKeepassDefaultGroup);
+		Assertions.assertEquals("Database non-existing.kdbx does not exist!", exception.getMessage());
+	}
+
+	@Test
+	void testExceptionWhenKeepassMasterKeyMissing() {
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY, "true");
+		final EnvConfigException exception = Assertions.assertThrows(EnvConfigException.class,
+				() -> EnvConfig.get("my.keepass.property"));
+		Assertions.assertEquals(String.format("Missing required variable '%s'", EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY),
+				exception.getMessage());
 	}
 
 	@Test
@@ -294,33 +345,11 @@ class EnvConfigTest {
 	}
 
 	@Test
-	void testCanGetEntryFromKeepassWhenFileNameSpecified() {
-		System.setProperty("config.keepass.filename", "env-config.kdbx");
-		enabledKeepass();
-		Assertions.assertEquals(KEEPASS_VALUE, EnvConfig.get("my.keepass.property"));
-	}
-
-	@Test
-	void testCanGetFromKeepassWhenFileNameWithSpace() {
-		System.setProperty("config.keepass.filename", "env config.kdbx");
-		enabledKeepass();
-		Assertions.assertEquals(KEEPASS_VALUE, EnvConfig.get("my.keepass.property"));
-	}
-
-	@Test
-	void testExceptionWhenKeepassFileMissing() {
-		System.setProperty("config.keepass.filename", "non-existing");
-		final ConfigException exception = Assertions.assertThrows(ConfigException.class,
-				this::testCanGetEntryFromKeepassDefaultGroup);
-		Assertions.assertEquals("Database non-existing.kdbx does not exist!", exception.getMessage());
-	}
-
-	@Test
 	void testCanGetEntryFromKeepassDBWhenMultipleEnvironment() {
-		System.setProperty(CONFIG_KEEPASS_ENABLED_KEY, "true");
-		System.setProperty(CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY, "true");
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
 		final String testEnv = "keepass";
-		System.setProperty(CONFIG_ENV_KEY, "test," + testEnv);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "test," + testEnv);
 		// when my.keepass.property exists in test env
 		// and only exists in keepass env group of keepass
 		// then keepass takes priority
@@ -330,12 +359,12 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetPropertyFromKeepassWhenMultipleEnv() {
-		environmentVariables.set(CONFIG_KEEPASS_ENABLED_KEY.replace(".", "_").toUpperCase(), "true");
-		environmentVariables.set(CONFIG_KEEPASS_MASTERKEY_KEY.replace(".", "_").toUpperCase(), CONFIG_KEEPASS_PASSWORD);
+		environmentVariables.set(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY.replace(".", "_").toUpperCase(), "true");
+		environmentVariables.set(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY.replace(".", "_").toUpperCase(), CONFIG_KEEPASS_PASSWORD);
 		environmentVariables.set("MY_KEEPASS_PROPERTY", "KEEPASS_ENV_VALUE");
 //		environmentVariables.set("PROPERTY_ONE", "KEEPASS_ENV");
 		final String testEnv = "test-no-keepass";
-		System.setProperty(CONFIG_ENV_KEY, "keepass," + testEnv);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, "keepass," + testEnv);
 		Assertions.assertEquals(testEnv, EnvConfig.getEnvironment());
 		// When a property.one exists in all environments, including default
 		// Then keepass takes priority
@@ -344,9 +373,9 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetEntryFromKeepassDB() {
-		System.setProperty(CONFIG_KEEPASS_ENABLED_KEY, "true");
-		System.setProperty(CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY, "true");
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		// when my.keepass.property exists in test env
 		// and only exists in default group of keepass
 		// then keepass takes priority
@@ -356,9 +385,9 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetEntryFromKeepassDisabled() {
-		System.setProperty(CONFIG_KEEPASS_ENABLED_KEY, "false");
-		System.setProperty(CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY, "false");
+		System.setProperty(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY, CONFIG_KEEPASS_PASSWORD);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		// when my.keepass.property exists in test env
 		// and only exists in default group of keepass
 		// then keepass takes priority
@@ -367,9 +396,9 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetKeepassOnlyEntry() {
-		environmentVariables.set(CONFIG_KEEPASS_ENABLED_KEY.replace(".", "_").toUpperCase(), "true");
-		environmentVariables.set(CONFIG_KEEPASS_MASTERKEY_KEY.replace(".", "_").toUpperCase(), CONFIG_KEEPASS_PASSWORD);
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		environmentVariables.set(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY.replace(".", "_").toUpperCase(), "true");
+		environmentVariables.set(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY.replace(".", "_").toUpperCase(), CONFIG_KEEPASS_PASSWORD);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		// when another.property does not exist in test env
 		// and exists in test group of keepass
 		Assertions.assertEquals("ANOTHER_PROPERTY", EnvConfig.get("another.property"));
@@ -377,9 +406,9 @@ class EnvConfigTest {
 
 	@Test
 	void testCanGetKeepassOnlyEntryWhenEntryWithTrailingSpace() {
-		environmentVariables.set(CONFIG_KEEPASS_ENABLED_KEY.replace(".", "_").toUpperCase(), "true");
-		environmentVariables.set(CONFIG_KEEPASS_MASTERKEY_KEY.replace(".", "_").toUpperCase(), CONFIG_KEEPASS_PASSWORD);
-		System.setProperty(CONFIG_ENV_KEY, TEST_ENVIRONMENT);
+		environmentVariables.set(EnvConfigProperties.CONFIG_KEEPASS_ENABLED_KEY.replace(".", "_").toUpperCase(), "true");
+		environmentVariables.set(EnvConfigProperties.CONFIG_KEEPASS_MASTERKEY_KEY.replace(".", "_").toUpperCase(), CONFIG_KEEPASS_PASSWORD);
+		System.setProperty(EnvConfigProperties.CONFIG_ENV_KEY, TEST_ENVIRONMENT);
 		// when another.property does not exist in test env
 		// and exists in test group of keepass
 		Assertions.assertEquals(KEEPASS_VALUE, EnvConfig.get("trailing.space.property"));